fix documentation

+++ .pre-commit-hooks/make-autodocs.py
@@ -13,6 +13,7 @@ loader_dirs = (
 for ldir in loader_dirs:
     autodocs[ldir] = []
 
+trans = str.maketrans({"_": r"\_"})
 docs_path = Path("docs")
 ref_path = docs_path / "ref"
 
@@ -22,21 +23,24 @@ for path in Path("src").glob("**/*.py"):
         import_path = ".".join(path.with_suffix("").parts[1:])
         autodocs[kind].append(import_path)
         rst_path = ref_path / kind / (import_path + ".rst")
+        if rst_path.is_file():
+            continue
         rst_path.parent.mkdir(parents=True, exist_ok=True)
         rst_path.write_text(
-            f"""
-{import_path}
-{'='*len(import_path)}
+            f"""{import_path.translate(trans)}
+{'='*len(import_path.translate(trans))}
 
-.. automodule:: {import_path}
-    :members:
-"""
+.. currentmodule:: {'.'.join(import_path.split('.')[:-1])}
+
+.. autodata:: {import_path.split('.')[-1]}"""
         )
 
 for ldir in autodocs:
     if not autodocs[ldir]:
         continue
     all_rst = ref_path / ldir / "all.rst"
+    if all_rst.is_file():
+        continue
     all_rst.parent.mkdir(parents=True, exist_ok=True)
     all_rst.write_text(
         f"""
